SeedDb

+++ SimplyForum/Controllers/CommentController.cs
@@ -1,10 +1,13 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity.UI.V4.Pages.Account.Internal;
+using Microsoft.AspNetCore.Mvc;
 using SimplyForum.Core.Contracts;
 using SimplyForum.Core.Models.Coment;
 using System.Security.Claims;
 
 namespace SimplyForum.Controllers
 {
+    [Authorize]
     public class CommentController : Controller
     {
         private readonly ICommentService commentService;
@@ -31,6 +34,13 @@ namespace SimplyForum.Controllers
         [HttpPost]
         public async Task<IActionResult> DeleteComment(DeleteCommentModel model)
         {
+            var currentAuthorId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value!;
+            var actualAuthorId = await commentService.GetCommentAuthorId(model.CommentId);
+            if (actualAuthorId != currentAuthorId || !User.IsInRole("Administrator"))
+            {
+                return Forbid();
+            
+            }
             await commentService.DeleteComment(model.CommentId);
 
             return RedirectToAction("Details", "Post", new { postId = model.PostId});
